+++ deps.ts
@@ -1,1 +1,0 @@
-export * as Console from "https://deno.land/std@0.87.0/fmt/colors.ts";