+++ src/bitarray/Write.ts
@@ -19,12 +19,12 @@ export class Write {
     }
 
     number(val: number, length: number) {
-        /*for (let i = length - 1; i >= 0; i--) {
-            this.bit((val & Math.pow(2, i)) >> i);
-        }*/
-        for (let i = 0; i < length; i++) {
+        for (let i = length - 1; i >= 0; i--) {
             this.bit((val & Math.pow(2, i)) >> i);
         }
+        /*for (let i = 0; i < length; i++) {
+            this.bit((val & Math.pow(2, i)) >> i);
+        }*/
     }
 
     uint8(val: number) {