+++ CHANGELOG.md
@@ -9,8 +9,11 @@
 ### Added
 
 -   Added `print` function that logs into console byteset and show current location
--   Added `isEnd` property to ByteSet
+-   Added `isEnd` property to `ByteSet`
 -   Added more Big Endian in functions
 -   Added `uint24` write and read for colors
 -   Added `each` function to go throuch each buffer element. You can set number type for example `uint8` for each byte or `uint16` for each short. By deafult it goes until position is reach end of the buffer.
 -   Added `limit` parameter to `each` function
+-   Added read `int32Array` and `uint32Array`
+-   Added `slice` for `ByteSet`
+-   Added `forEach` for `ByteSet`